Move duplicated code to common helper
